Slot_Attention_with_Prototypes/requirements.txt
pytorch
h5py
matplotlib
tensorboard
torchsummary
pycocotools
scipy
scikit-learn
captum
setproctitle